Inline script_pubkey Script extension traits
The separation of extension traits into the `script_pubkey` module
does not assist clarity and makes it less ergonomic for users having
to import traits from different locations. We can put these methods on
the `script::ScriptExt` and fix both issues.
