+++ bitcoin/src/address/script_pubkey.rs
@@ -2,7 +2,6 @@
 
 //! Bitcoin scriptPubkey script extensions.
 
-use internals::array::ArrayExt;
 use secp256k1::{Secp256k1, Verification};
 
 use crate::internal_macros::define_extension_trait;
@@ -12,89 +11,8 @@ use crate::key::{
 use crate::opcodes::all::*;
 use crate::script::witness_program::{WitnessProgram, P2A_PROGRAM};
 use crate::script::witness_version::WitnessVersion;
-use crate::script::{
-    self, Builder, PushBytes, RedeemScriptSizeError, Script, ScriptBuf, ScriptExt as _, ScriptHash,
-    WScriptHash, WitnessScriptSizeError,
-};
+use crate::script::{Builder, PushBytes, Script, ScriptBuf, ScriptHash, WScriptHash};
 use crate::taproot::TapNodeHash;
-
-define_extension_trait! {
-    /// Extension functionality to add scriptPubkey support to the [`Script`] type.
-    pub trait ScriptExt impl for Script {
-        /// Computes the P2WSH output corresponding to this witnessScript (aka the "witness redeem
-        /// script").
-        fn to_p2wsh(&self) -> Result<ScriptBuf, WitnessScriptSizeError> {
-            self.wscript_hash().map(ScriptBuf::new_p2wsh)
-        }
-
-        /// Computes P2TR output with a given internal key and a single script spending path equal to
-        /// the current script, assuming that the script is a Tapscript.
-        fn to_p2tr<C: Verification, K: Into<UntweakedPublicKey>>(
-            &self,
-            secp: &Secp256k1<C>,
-            internal_key: K,
-        ) -> ScriptBuf {
-            let internal_key = internal_key.into();
-            let leaf_hash = self.tapscript_leaf_hash();
-            let merkle_root = TapNodeHash::from(leaf_hash);
-            ScriptBuf::new_p2tr(secp, internal_key, Some(merkle_root))
-        }
-
-        /// Computes the P2SH output corresponding to this redeem script.
-        fn to_p2sh(&self) -> Result<ScriptBuf, RedeemScriptSizeError> {
-            self.script_hash().map(ScriptBuf::new_p2sh)
-        }
-
-        /// Returns the script code used for spending a P2WPKH output if this script is a script pubkey
-        /// for a P2WPKH output. The `scriptCode` is described in [BIP143].
-        ///
-        /// [BIP143]: <https://github.com/bitcoin/bips/blob/99701f68a88ce33b2d0838eb84e115cef505b4c2/bip-0143.mediawiki>
-        fn p2wpkh_script_code(&self) -> Option<ScriptBuf> {
-            if self.is_p2wpkh() {
-                // The `self` script is 0x00, 0x14, <pubkey_hash>
-                let bytes = <[u8; 20]>::try_from(&self.as_bytes()[2..]).expect("length checked in is_p2wpkh()");
-                let wpkh = WPubkeyHash::from_byte_array(bytes);
-                Some(script::p2wpkh_script_code(wpkh))
-            } else {
-                None
-            }
-        }
-
-        /// Checks whether a script pubkey is a P2PK output.
-        ///
-        /// You can obtain the public key, if its valid,
-        /// by calling [`p2pk_public_key()`](Self::p2pk_public_key)
-        fn is_p2pk(&self) -> bool { self.p2pk_pubkey_bytes().is_some() }
-
-        /// Returns the public key if this script is P2PK with a **valid** public key.
-        ///
-        /// This may return `None` even when [`is_p2pk()`](Self::is_p2pk) returns true.
-        /// This happens when the public key is invalid (e.g. the point not being on the curve).
-        /// In this situation the script is unspendable.
-        fn p2pk_public_key(&self) -> Option<PublicKey> {
-            PublicKey::from_slice(self.p2pk_pubkey_bytes()?).ok()
-        }
-    }
-}
-
-define_extension_trait! {
-    pub(crate) trait ScriptExtPrivate impl for Script {
-        /// Returns the bytes of the (possibly invalid) public key if this script is P2PK.
-        fn p2pk_pubkey_bytes(&self) -> Option<&[u8]> {
-            if let Ok(bytes) = <&[u8; 67]>::try_from(self.as_bytes()) {
-                let (&first, bytes) = bytes.split_first::<66>();
-                let (&last, pubkey) = bytes.split_last::<65>();
-                (first == OP_PUSHBYTES_65.to_u8() && last == OP_CHECKSIG.to_u8()).then_some(pubkey)
-            } else if let Ok(bytes) = <&[u8; 35]>::try_from(self.as_bytes()) {
-                let (&first, bytes) = bytes.split_first::<34>();
-                let (&last, pubkey) = bytes.split_last::<33>();
-                (first == OP_PUSHBYTES_33.to_u8() && last == OP_CHECKSIG.to_u8()).then_some(pubkey)
-            } else {
-                None
-            }
-        }
-    }
-}
 
 define_extension_trait! {
     /// Extension functionality to add scriptPubkey support to the [`ScriptBuf`] type.
@@ -195,6 +113,7 @@ mod sealed {
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::script::ScriptExt as _;
 
     #[test]
     fn shortest_witness_program() {
